\name{paleoCLMs-package}
\alias{paleoCLMs-package}
\alias{paleoCLMs}
\docType{package}
\title{
Miscelanea functions for the paleoCLM project}
\description{
This package provide a set of handy functions for the paleoCLM project. This functions cover from loading data to fit, evaluate, and compare models.}
\details{
\tabular{ll}{
Package: \tab paleoCLMs\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2014-02-21\cr
License: \tab What license is it under?\cr
}
}
\author{
Diego Nieto Lugilde

Maintainer: Diego Nieto Lugilde <dinilu@ugr.es>
}
\keyword{ package }
